requirements.txt
# Restoration-Intel Python Dependencies

# Web Framework and Server
fastapi==0.95.2
uvicorn==0.22.0
httpx==0.24.1

# Database and ORM
sqlalchemy==1.4.41
psycopg2-binary==2.9.6
python-dotenv==1.0.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Numerical and Scientific Computing
numpy>=1.22.0,<1.24.0  # Flexible version to satisfy multiple dependencies
pandas==2.0.1
scipy==1.10.1
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2

# Machine Learning
tensorflow-cpu==2.12.0
xgboost==1.7.5

# Probabilistic and Statistical Modeling
statsmodels==0.14.0

# Time Series and Forecasting
prophet==1.1.2
pmdarima==2.0.1

# Plotting and Visualization
plotly==5.14.1
bokeh==2.4.3

# Logging and Monitoring
loguru==0.7.0

# Scheduling and Background Tasks
schedule==1.2.0

# Development and Debugging
ipython==8.12.0
pytest==7.3.1

# Optional: Additional Utilities
python-multipart==0.0.6
pydantic==1.10.7